Add support for writing dmcrypt'd partitions
This only adds support for the aes-cbc-plain mode for
dmcrypt.

The raw_write function now has an optional parameter for the write mode
and an option. Pass "aes-cbc-plain" to raw_write and the secret. While
the secret is shown hardcoded into the fwup.conf file, the likely usage
is to pass it in via a variable so that it can be stored
outside of version control or set specific to a device.