Replace libsodium with Monocypher
Monocypher (https://monocypher.org/) is an easy-to-use crypto library
that's similar to libsodium, but has a much smaller footprint. It
supports the same algorithms so it's nearly a 1:1 replacement.

A static x86_64 build of fwup with Monocypher is about 25% smaller than
with libsodium (saves 164 KB).

On shared library builds of fwup, it is frequently the case that fwup is
the only library on an embedded device. Removing libsodium.so is a
reduction of 260-300 KB in exchange for <50 KB of code (based on
the size of monocypher.o).